use BF16 for training

+++ mlm_pretrain_v3.py
@@ -380,8 +380,8 @@ def train_mlm_multi_gpu(model, train_loader, val_loader, device, tokenizer, num_
             # Clear gradients
             optimizer.zero_grad()
             
-            # Forward pass with mixed precision
-            with autocast():
+            # Forward pass with bfloat16 mixed precision (better for H200)
+            with autocast(dtype=torch.bfloat16):
                 loss = model(input_ids, labels)
             
             # Backward pass with gradient scaling
@@ -416,7 +416,7 @@ def train_mlm_multi_gpu(model, train_loader, val_loader, device, tokenizer, num_
                 input_ids = batch['input_ids'].to(device)
                 labels = batch['labels'].to(device)
                 
-                with autocast():
+                with autocast(dtype=torch.bfloat16):
                     loss = model(input_ids, labels)
                 
                 total_val_loss += loss.item()
